requirements.txt
prompt_toolkit
azure-mgmt-resource~=1.1.0
azure-mgmt-network~=1.0.0
azure-mgmt-compute~=2.0.0
haikunator
